Implemented the adjusted matching L & R gait cycles, and also the non-matching gait cycle segmentation
Segmented the matching gait cycles using each set of pairs, not skipping any, per the discussion in the meeting today (previously I was skipping the alternating pairs of gait cycles).

When I implemented this, I realized that the way I was storing all of the data would now result in duplicate data being in the exported CSV file, which is not desired. So I created splitTrialsByGaitCycle_NoMatching.m because to store the non-symmetry values I need a non-matching data table.

The symmetry values will then be in their own separate table.