Uploading CGAM Code + ML Code
CGAM function now updated to take in a string to designate which features are to be included in the CGAM calculation. Note: This will only include the designated features selected after the initial filter/removal of features. For example: 'emg' will only have the peak index sym in the CGAM calc, as all other EMG features were dropped prior.